move `filter_ZnTrackOption` to utils
